requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-aws==0.1.5
neo4j==5.20.0
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.35.0
boto3==1.34.108
sagemaker==2.219.0
tiktoken==0.7.0